Use RDS broker via the ELB.
This increases the broker instance count to 2, and adds the instances to
the ELB. It also updates the broker registration to point at the ELB
instead of directly at the static IP of the instance.

Now that we're using an ELB, there's no longer a need for the broker
instances to have static IPs.