Reduce size of rds_broker VMs to t2.nano
These are currently way overspeccd for what's needed. The broker itself
is a small Go application that doesn't handle many requests - it's only
called when a user creates/deletes or binds/unbinds a service instance.

Current usage stats for an instance that's been running for a while:

```
$ free
             total       used       free     shared    buffers     cached
Mem:       3853840     206868    3646972        372      13120     117332
-/+ buffers/cache:      76416    3777424
Swap:      3855592          0    3855592

$ uptime
 10:49:20 up 58 min,  1 user,  load average: 0.00, 0.01, 0.05

$ mpstat
Linux 3.19.0-59-generic (194770cf-6911-489e-83ae-ae2773a4460e)  06/30/2016  _x86_64_  (1 CPU)

10:48:33 AM  CPU    %usr   %nice    %sys %iowait    %irq   %soft  %steal  %guest  %gnice   %idle
10:48:33 AM  all    0.22    0.00    0.25    1.09    0.00    0.00    0.35    0.00    0.00   98.08
```

This shows that only 200M of RAM is used (including disk cache), and that the
CPU usage is minimal. During a run of the custom acceptance tests, I was
the CPU usage peak at 2% while running mpstate with a 1 second interval.

It's therefore reasonable to change the instance size to a t2.nano. This
comes with 0.5G RAM, and a single CPU with burstable performance[1]. The
t2.nano comes with 3 CPU credits per hour, which works out at a baseline
performance of 5% of a CPU core, with the ability to burst to 100% for
short periods[2].

[1]https://aws.amazon.com/ec2/instance-types/#t2
[2]https://aws.amazon.com/ec2/instance-types/#burst