Refactored generation of test configuration
The task for generating the test configuration file for the
custom acceptance tests and the performance tests have been
moved to an external task file. This is because the configurations
are exactly the same, therefore duplicating source code.

This external task file can now be referenced by a failure test,
which will be introduced to test the now highly available RDS
broker.