Added failure test for RDS broker
The failure test destroys `rds_broker/0`, which prior to making
the broker highly available was the only broker instance. It then
runs the custom acceptance tests, but uses `ginkgo.focus` to run
only the RDS broker test.

The `ginkgo.focus` parameter is passed in via an environment variable,
which will make it easier to focus tests in future.