Import Grafana dashboards into cf-manifest
The script takes all JSON files, minifies them, then constructs
CloudFoundry manifest properties in YAML and sends it to standard
output. This will allow us to create YAML files in the Concourse
pipeline, which can be merged into the CloudFoundry manifest with
Spruce.

Minifying the JSON was chosen to avoid converting to YAML, as we
would just have to convert it back to JSON for use in the Grafana
bosh release.

The task for creating the Grafana manifest properties has been
aggregated with the previous task (extract-terraform-outputs) as
they are independent of one another. This will reduce the amount of
time this extra task adds to the job.